% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarSdSig.R
\name{plotBarSdSig}
\alias{plotBarSdSig}
\title{plotBarSdSig}
\usage{
plotBarSdSig(data, x, y, rowOrder = c(), comparisons = NULL, y_position = NULL)
}
\arguments{
\item{data}{Default dataset to use for plot.}

\item{x}{Default variable from data for aesthetic mappings to use for plot}

\item{y}{Default variable from data for aesthetic mappings to use for plot}

\item{rowOrder}{the order of x-axis}

\item{comparisons}{A list of length-2 vectors. The entries in the vector are either the names of 2 values on the x-axis or the 2 integers that correspond to the index of the columns of interest.}

\item{y_position}{numeric vector with the y positions of the brackets}
}
\value{
a ggplot2 object
}
\description{
This function would generate a list including a countData and colData as input data for testing
}
\examples{
\dontrun{
library(ggplot2);library(dplyr);library(ggsignif)

compareGroup = list(c("f","4"),c("f","r"),c("4","r"))
plotBarSdSig(mpg,drv,hwy,comparisons=compareGroup,y_position=TRUE)

plotBarSdSig(mpg,drv,hwy,rowOrder=c("r","4","f"),
comparisons=compareGroup,y_position=c(48.2,50.8,32.07))
}
}
