% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLmCor.R
\name{plotLmCor}
\alias{plotLmCor}
\title{plotLmCor}
\usage{
plotLmCor(mydata)
}
\arguments{
\item{mydata}{a dataframe with two column of x and y}
}
\value{
a ggplot2 object
}
\description{
This function would generate a list including a countData and colData as input data for testing
}
\examples{
mydata = mpg[,8:9];colnames(mydata)=c("x","y")
plotLmCor(mydata)
}
