% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLmCor.R
\name{plotLmCor2}
\alias{plotLmCor2}
\title{plotLmCor2}
\usage{
plotLmCor2(data, x, y)
}
\arguments{
\item{data}{Default dataset to use for plot.}

\item{x}{Default variable from data for aesthetic mappings to use for plot}

\item{y}{Default variable from data for aesthetic mappings to use for plot}
}
\value{
a ggplot2 object
}
\description{
This function would generate a list including a countData and colData as input data for testing
}
\examples{
require(ggplot2)
print(plotLmCor2(mpg,cty,hwy))
}
