% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarSdSig.R
\name{plotBarSdSig}
\alias{plotBarSdSig}
\title{plotBarSdSig}
\usage{
plotBarSdSig(mydata)
}
\arguments{
\item{mydata}{a dataframe with two column of group and phe}
}
\value{
a ggplot2 object
}
\description{
This function would generate a list including a countData and colData as input data for testing
}
\examples{
mydata = mpg[,c("drv","hwy")];colnames(mydata)=c("group","phe")
plotBarSdSig(mydata)
}
