% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLmCor.R
\name{plotLmCor}
\alias{plotLmCor}
\title{plotLmCor}
\usage{
plotLmCor(data, mapping)
}
\arguments{
\item{mydata}{a dataframe with two column of x and y}
}
\value{
a ggplot2 object
}
\description{
This function would generate a list including a countData and colData as input data for testing
}
\examples{
require(ggplot2)
print(plotLmCor(mpg,aes(cty,hwy)))
}
